refactor(global): 순서 검증 관련 클래스 패키지 변경
