+++ backend/src/main/java/codezap/template/dto/request/CreateTemplateRequest.java
@@ -11,7 +11,7 @@ import codezap.global.validation.ByteLength;
 import codezap.global.validation.ValidationGroups.NotNullGroup;
 import codezap.global.validation.ValidationGroups.SizeCheckGroup;
 import codezap.template.domain.Visibility;
-import codezap.template.dto.request.validation.ValidatedOrdinalRequest;
+import codezap.global.validation.ValidatedOrdinalRequest;
 import codezap.template.dto.request.validation.ValidatedSourceCodesCountRequest;
 import io.swagger.v3.oas.annotations.media.Schema;
