+++ backend/src/main/java/codezap/template/service/SourceCodeService.java
@@ -16,7 +16,7 @@ import codezap.template.dto.request.CreateSourceCodeRequest;
 import codezap.template.dto.request.CreateTemplateRequest;
 import codezap.template.dto.request.UpdateSourceCodeRequest;
 import codezap.template.dto.request.UpdateTemplateRequest;
-import codezap.template.dto.request.validation.ValidatedOrdinalRequest;
+import codezap.global.validation.ValidatedOrdinalRequest;
 import codezap.template.dto.request.validation.ValidatedSourceCodesCountRequest;
 import codezap.template.repository.SourceCodeRepository;
 import lombok.RequiredArgsConstructor;
